+++ pkg/utils/mock.go
@@ -10,7 +10,7 @@ import (
 func MockComputeEngineInsights() *pb.Entity {
 	entity := pb.Entity{
 		Id:          "computeEngine",
-		Aggregation: []int32{80000, 90000},
+		Aggregation: []float64{80000, 90000},
 		Change: &pb.ChangeStatistic{
 			Ratio:  0.125,
 			Amount: 10000,
@@ -19,7 +19,7 @@ func MockComputeEngineInsights() *pb.Entity {
 			Service: []*pb.Entity{
 				&pb.Entity{
 					Id:          "service-a",
-					Aggregation: []int32{20000, 10000},
+					Aggregation: []float64{20000, 10000},
 					Change: &pb.ChangeStatistic{
 						Ratio:  -0.5,
 						Amount: -10000,
@@ -28,7 +28,7 @@ func MockComputeEngineInsights() *pb.Entity {
 						SKU: []*pb.Entity{
 							&pb.Entity{
 								Id:          "Mock SKU A",
-								Aggregation: []int32{4000, 2000},
+								Aggregation: []float64{4000, 2000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  -0.5,
 									Amount: -2000,
@@ -37,7 +37,7 @@ func MockComputeEngineInsights() *pb.Entity {
 							},
 							&pb.Entity{
 								Id:          "Mock SKU B",
-								Aggregation: []int32{7000, 6000},
+								Aggregation: []float64{7000, 6000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  -0.14285714285714285,
 									Amount: -1000,
@@ -46,7 +46,7 @@ func MockComputeEngineInsights() *pb.Entity {
 							},
 							&pb.Entity{
 								Id:          "Mock SKU C",
-								Aggregation: []int32{9000, 2000},
+								Aggregation: []float64{9000, 2000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  -0.7777777777777778,
 									Amount: -7000,
@@ -57,7 +57,7 @@ func MockComputeEngineInsights() *pb.Entity {
 						Deployment: []*pb.Entity{
 							&pb.Entity{
 								Id:          "Compute Engine",
-								Aggregation: []int32{7000, 6000},
+								Aggregation: []float64{7000, 6000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  -0.5,
 									Amount: -2000,
@@ -66,7 +66,7 @@ func MockComputeEngineInsights() *pb.Entity {
 							},
 							&pb.Entity{
 								Id:          "Kubernetes",
-								Aggregation: []int32{4000, 2000},
+								Aggregation: []float64{4000, 2000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  -0.14285714285714285,
 									Amount: -1000,
@@ -79,7 +79,7 @@ func MockComputeEngineInsights() *pb.Entity {
 				},
 				&pb.Entity{
 					Id:          "service-b",
-					Aggregation: []int32{10000, 20000},
+					Aggregation: []float64{10000, 20000},
 					Change: &pb.ChangeStatistic{
 						Ratio:  1,
 						Amount: 10000,
@@ -88,7 +88,7 @@ func MockComputeEngineInsights() *pb.Entity {
 						SKU: []*pb.Entity{
 							&pb.Entity{
 								Id:          "Mock SKU A",
-								Aggregation: []int32{1000, 2000},
+								Aggregation: []float64{1000, 2000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  1,
 									Amount: 1000,
@@ -97,7 +97,7 @@ func MockComputeEngineInsights() *pb.Entity {
 							},
 							&pb.Entity{
 								Id:          "Mock SKU B",
-								Aggregation: []int32{4000, 8000},
+								Aggregation: []float64{4000, 8000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  1,
 									Amount: 4000,
@@ -106,7 +106,7 @@ func MockComputeEngineInsights() *pb.Entity {
 							},
 							&pb.Entity{
 								Id:          "Mock SKU C",
-								Aggregation: []int32{5000, 10000},
+								Aggregation: []float64{5000, 10000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  1,
 									Amount: 5000,
@@ -117,7 +117,7 @@ func MockComputeEngineInsights() *pb.Entity {
 						Deployment: []*pb.Entity{
 							&pb.Entity{
 								Id:          "Compute Engine",
-								Aggregation: []int32{7000, 6000},
+								Aggregation: []float64{7000, 6000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  -0.5,
 									Amount: -2000,
@@ -126,7 +126,7 @@ func MockComputeEngineInsights() *pb.Entity {
 							},
 							&pb.Entity{
 								Id:          "Kubernetes",
-								Aggregation: []int32{4000, 2000},
+								Aggregation: []float64{4000, 2000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  -0.14285714285714285,
 									Amount: -1000,
@@ -139,7 +139,7 @@ func MockComputeEngineInsights() *pb.Entity {
 				},
 				&pb.Entity{
 					Id:          "service-c",
-					Aggregation: []int32{0, 10000},
+					Aggregation: []float64{0, 10000},
 					Change: &pb.ChangeStatistic{
 						Amount: 10000,
 					},
@@ -155,7 +155,7 @@ func MockComputeEngineInsights() *pb.Entity {
 func MockCloudDataflowInsights() *pb.Entity{
 	entity := pb.Entity{
 		Id:          "cloudDataflow",
-		Aggregation: []int32{100000, 158000},
+		Aggregation: []float64{100000, 158000},
 		Change: &pb.ChangeStatistic{
 			Ratio:  0.58,
 			Amount: 50000,
@@ -163,7 +163,7 @@ func MockCloudDataflowInsights() *pb.Entity{
 		Entities: &pb.Record{
 			Pipeline: []*pb.Entity{
 				&pb.Entity{
-					Aggregation: []int32{10000, 12000},
+					Aggregation: []float64{10000, 12000},
 					Change: &pb.ChangeStatistic{
 						Ratio:  0.2,
 						Amount: -2000,
@@ -172,7 +172,7 @@ func MockCloudDataflowInsights() *pb.Entity{
 						SKU: []*pb.Entity{
 							&pb.Entity{
 								Id:          "Mock SKU A",
-								Aggregation: []int32{3000, 4000},
+								Aggregation: []float64{3000, 4000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  0.333333,
 									Amount: 12000,
@@ -181,7 +181,7 @@ func MockCloudDataflowInsights() *pb.Entity{
 							},
 							&pb.Entity{
 								Id:          "Mock SKU B",
-								Aggregation: []int32{7000, 8000},
+								Aggregation: []float64{7000, 8000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  0.14285714,
 									Amount: 1000,
@@ -192,7 +192,7 @@ func MockCloudDataflowInsights() *pb.Entity{
 						Deployment: []*pb.Entity{
 							&pb.Entity{
 								Id:          "Compute Engine",
-								Aggregation: []int32{7000, 6000},
+								Aggregation: []float64{7000, 6000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  -0.5,
 									Amount: -2000,
@@ -201,7 +201,7 @@ func MockCloudDataflowInsights() *pb.Entity{
 							},
 							&pb.Entity{
 								Id:          "Kubernetes",
-								Aggregation: []int32{4000, 2000},
+								Aggregation: []float64{4000, 2000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  -0.14285714285714285,
 									Amount: -1000,
@@ -214,7 +214,7 @@ func MockCloudDataflowInsights() *pb.Entity{
 				},
 				&pb.Entity{
 					Id: "pipeline-a",
-					Aggregation: []int32{60000, 70000},
+					Aggregation: []float64{60000, 70000},
 					Change: &pb.ChangeStatistic{
 						Ratio:  0.16666666666666666,
 						Amount: 10000,
@@ -223,7 +223,7 @@ func MockCloudDataflowInsights() *pb.Entity{
 						SKU: []*pb.Entity{
 							&pb.Entity{
 								Id:          "Mock SKU A",
-								Aggregation: []int32{20000, 15000},
+								Aggregation: []float64{20000, 15000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  -0.25,
 									Amount: -5000,
@@ -232,7 +232,7 @@ func MockCloudDataflowInsights() *pb.Entity{
 							},
 							&pb.Entity{
 								Id:          "Mock SKU B",
-								Aggregation: []int32{30000, 35000},
+								Aggregation: []float64{30000, 35000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  -0.16666666666666666,
 									Amount: -5000,
@@ -241,7 +241,7 @@ func MockCloudDataflowInsights() *pb.Entity{
 							},
 							&pb.Entity{
 								Id:          "Mock SKU C",
-								Aggregation: []int32{10000, 20000},
+								Aggregation: []float64{10000, 20000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  1,
 									Amount: 10000,
@@ -252,7 +252,7 @@ func MockCloudDataflowInsights() *pb.Entity{
 						Deployment: []*pb.Entity{
 							&pb.Entity{
 								Id:          "Compute Engine",
-								Aggregation: []int32{7000, 6000},
+								Aggregation: []float64{7000, 6000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  -0.5,
 									Amount: -2000,
@@ -261,7 +261,7 @@ func MockCloudDataflowInsights() *pb.Entity{
 							},
 							&pb.Entity{
 								Id:          "Kubernetes",
-								Aggregation: []int32{4000, 2000},
+								Aggregation: []float64{4000, 2000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  -0.14285714285714285,
 									Amount: -1000,
@@ -274,7 +274,7 @@ func MockCloudDataflowInsights() *pb.Entity{
 				},
 				&pb.Entity{
 					Id: "pipeline-b",
-					Aggregation: []int32{12000, 8000},
+					Aggregation: []float64{12000, 8000},
 					Change: &pb.ChangeStatistic{
 						Ratio:  -0.33333,
 						Amount: -4000,
@@ -283,7 +283,7 @@ func MockCloudDataflowInsights() *pb.Entity{
 						SKU: []*pb.Entity{
 							&pb.Entity{
 								Id:          "Mock SKU A",
-								Aggregation: []int32{4000, 4000},
+								Aggregation: []float64{4000, 4000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  0,
 									Amount: 0,
@@ -292,7 +292,7 @@ func MockCloudDataflowInsights() *pb.Entity{
 							},
 							&pb.Entity{
 								Id:          "Mock SKU B",
-								Aggregation: []int32{8000, 4000},
+								Aggregation: []float64{8000, 4000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  -0.5,
 									Amount: -4000,
@@ -305,7 +305,7 @@ func MockCloudDataflowInsights() *pb.Entity{
 				},
 				&pb.Entity{
 					Id: "pipeline-c",
-					Aggregation: []int32{0, 10000},
+					Aggregation: []float64{0, 10000},
 					Change: &pb.ChangeStatistic{
 						Amount: 10000,
 					},
@@ -320,7 +320,7 @@ func MockCloudDataflowInsights() *pb.Entity{
 func MockCloudStorageInsights() *pb.Entity{
 	entity := pb.Entity{
 		Id:          "cloudStorage",
-		Aggregation: []int32{45000, 45000},
+		Aggregation: []float64{45000, 45000},
 		Change: &pb.ChangeStatistic{
 			Ratio:  0,
 			Amount: 0,
@@ -329,7 +329,7 @@ func MockCloudStorageInsights() *pb.Entity{
 			Bucket: []*pb.Entity{
 				&pb.Entity{
 					Id: "bucket-a",
-					Aggregation: []int32{15000, 20000},
+					Aggregation: []float64{15000, 20000},
 					Change: &pb.ChangeStatistic{
 						Ratio:  0.333,
 						Amount: 5000,
@@ -338,7 +338,7 @@ func MockCloudStorageInsights() *pb.Entity{
 						SKU: []*pb.Entity{
 							&pb.Entity{
 								Id:          "Mock SKU A",
-								Aggregation: []int32{10000, 11000},
+								Aggregation: []float64{10000, 11000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  0.1,
 									Amount: 1000,
@@ -347,7 +347,7 @@ func MockCloudStorageInsights() *pb.Entity{
 							},
 							&pb.Entity{
 								Id:          "Mock SKU B",
-								Aggregation: []int32{2000, 5000},
+								Aggregation: []float64{2000, 5000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  1.5,
 									Amount: 3000,
@@ -356,7 +356,7 @@ func MockCloudStorageInsights() *pb.Entity{
 							},
 							&pb.Entity{
 								Id:          "Mock SKU C",
-								Aggregation: []int32{3000, 4000},
+								Aggregation: []float64{3000, 4000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  0.3333,
 									Amount: 1000,
@@ -367,7 +367,7 @@ func MockCloudStorageInsights() *pb.Entity{
 						Deployment: []*pb.Entity{
 							&pb.Entity{
 								Id:          "Compute Engine",
-								Aggregation: []int32{7000, 6000},
+								Aggregation: []float64{7000, 6000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  -0.5,
 									Amount: -2000,
@@ -376,7 +376,7 @@ func MockCloudStorageInsights() *pb.Entity{
 							},
 							&pb.Entity{
 								Id:          "Kubernetes",
-								Aggregation: []int32{4000, 2000},
+								Aggregation: []float64{4000, 2000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  -0.14285714285714285,
 									Amount: -1000,
@@ -389,7 +389,7 @@ func MockCloudStorageInsights() *pb.Entity{
 				},
 				&pb.Entity{
 					Id: "bucket-b",
-					Aggregation: []int32{30000, 25000},
+					Aggregation: []float64{30000, 25000},
 					Change: &pb.ChangeStatistic{
 						Ratio:  -0.16666,
 						Amount: -5000,
@@ -398,7 +398,7 @@ func MockCloudStorageInsights() *pb.Entity{
 						SKU: []*pb.Entity{
 							&pb.Entity{
 								Id:          "Mock SKU A",
-								Aggregation: []int32{12000, 13000},
+								Aggregation: []float64{12000, 13000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  0.08333333333333333,
 									Amount: 1000,
@@ -407,7 +407,7 @@ func MockCloudStorageInsights() *pb.Entity{
 							},
 							&pb.Entity{
 								Id:          "Mock SKU B",
-								Aggregation: []int32{16000, 12000},
+								Aggregation: []float64{16000, 12000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  -0.25,
 									Amount: -4000,
@@ -416,7 +416,7 @@ func MockCloudStorageInsights() *pb.Entity{
 							},
 							&pb.Entity{
 								Id:          "Mock SKU C",
-								Aggregation: []int32{2000, 0},
+								Aggregation: []float64{2000, 0},
 								Change: &pb.ChangeStatistic{
 									Amount: -2000,
 								},
@@ -428,7 +428,7 @@ func MockCloudStorageInsights() *pb.Entity{
 				},
 				&pb.Entity{
 					Id: "bucket-c",
-					Aggregation: []int32{0, 0},
+					Aggregation: []float64{0, 0},
 					Change: &pb.ChangeStatistic{
 						Amount: 0,
 					},
@@ -443,7 +443,7 @@ func MockCloudStorageInsights() *pb.Entity{
 func MockBigQueryInsights() *pb.Entity{
 	entity := pb.Entity{
 		Id:          "bigQuery",
-		Aggregation: []int32{10000, 30000},
+		Aggregation: []float64{10000, 30000},
 		Change: &pb.ChangeStatistic{
 			Ratio:  3,
 			Amount: 20000,
@@ -452,7 +452,7 @@ func MockBigQueryInsights() *pb.Entity{
 			Dataset: []*pb.Entity{
 				&pb.Entity{
 					Id: "dataset-a",
-					Aggregation: []int32{5000, 10000},
+					Aggregation: []float64{5000, 10000},
 					Change: &pb.ChangeStatistic{
 						Ratio:  1,
 						Amount: 5000,
@@ -461,7 +461,7 @@ func MockBigQueryInsights() *pb.Entity{
 				},
 				&pb.Entity{
 					Id: "dataset-b",
-					Aggregation: []int32{5000, 10000},
+					Aggregation: []float64{5000, 10000},
 					Change: &pb.ChangeStatistic{
 						Ratio:  1,
 						Amount: 5000,
@@ -470,7 +470,7 @@ func MockBigQueryInsights() *pb.Entity{
 				},
 				&pb.Entity{
 					Id: "dataset-c",
-					Aggregation: []int32{0, 10000},
+					Aggregation: []float64{0, 10000},
 					Change: &pb.ChangeStatistic{
 						Amount: 10000,
 					},
@@ -485,7 +485,7 @@ func MockBigQueryInsights() *pb.Entity{
 func MockEventsInsights() *pb.Entity{
 	entity := pb.Entity{
 		Id:          "events",
-		Aggregation: []int32{20000, 10000},
+		Aggregation: []float64{20000, 10000},
 		Change: &pb.ChangeStatistic{
 			Ratio:  -0.5,
 			Amount: -10000,
@@ -494,7 +494,7 @@ func MockEventsInsights() *pb.Entity{
 			Event: []*pb.Entity{
 				&pb.Entity{
 					Id:          "event-a",
-					Aggregation: []int32{15000, 7000},
+					Aggregation: []float64{15000, 7000},
 					Change: &pb.ChangeStatistic{
 						Ratio:  -0.53333333333,
 						Amount: -8000,
@@ -503,7 +503,7 @@ func MockEventsInsights() *pb.Entity{
 						Product: []*pb.Entity{
 							&pb.Entity{
 								Id:          "Mock Product A",
-								Aggregation: []int32{5000, 2000},
+								Aggregation: []float64{5000, 2000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  -0.6,
 									Amount: -3000,
@@ -512,7 +512,7 @@ func MockEventsInsights() *pb.Entity{
 							},
 							&pb.Entity{
 								Id:          "Mock Product B",
-								Aggregation: []int32{7000, 2500},
+								Aggregation: []float64{7000, 2500},
 								Change: &pb.ChangeStatistic{
 									Ratio:  -0.64285714285,
 									Amount: -4500,
@@ -521,7 +521,7 @@ func MockEventsInsights() *pb.Entity{
 							},
 							&pb.Entity{
 								Id:          "Mock Product C",
-								Aggregation: []int32{3000, 2500},
+								Aggregation: []float64{3000, 2500},
 								Change: &pb.ChangeStatistic{
 									Ratio:  -0.16666666666,
 									Amount: -500,
@@ -533,7 +533,7 @@ func MockEventsInsights() *pb.Entity{
 				},
 				&pb.Entity{
 					Id:          "event-b",
-					Aggregation: []int32{5000, 3000},
+					Aggregation: []float64{5000, 3000},
 					Change: &pb.ChangeStatistic{
 						Ratio:  -0.4,
 						Amount: -2000,
@@ -542,7 +542,7 @@ func MockEventsInsights() *pb.Entity{
 						Product: []*pb.Entity{
 							&pb.Entity{
 								Id:          "Mock Product A",
-								Aggregation: []int32{2000, 1000},
+								Aggregation: []float64{2000, 1000},
 								Change: &pb.ChangeStatistic{
 									Ratio:  -0.5,
 									Amount: -1000,
@@ -551,7 +551,7 @@ func MockEventsInsights() *pb.Entity{
 							},
 							&pb.Entity{
 								Id:          "Mock Product B",
-								Aggregation: []int32{1000, 1500},
+								Aggregation: []float64{1000, 1500},
 								Change: &pb.ChangeStatistic{
 									Ratio:  -0.5,
 									Amount: 500,
@@ -560,7 +560,7 @@ func MockEventsInsights() *pb.Entity{
 							},
 							&pb.Entity{
 								Id:          "Mock Product C",
-								Aggregation: []int32{2000, 500},
+								Aggregation: []float64{2000, 500},
 								Change: &pb.ChangeStatistic{
 									Ratio:  -0.75,
 									Amount: -1500,
@@ -583,7 +583,7 @@ func MockAlerts() []*pb.Entity{
 		Project:          "example-project",
 		PeriodStart: "2020-02",
 		PeriodEnd: "2020-03",
-		Aggregation: []int32{60000, 120000},
+		Aggregation: []float64{60000, 120000},
 		Change: &pb.ChangeStatistic{
 			Ratio:  1,
 			Amount: 60000,
@@ -591,15 +591,15 @@ func MockAlerts() []*pb.Entity{
 		Products: []*pb.Entity{
 			&pb.Entity{
 				Id:          "Compute Engine",
-				Aggregation: []int32{50000, 118000},
+				Aggregation: []float64{50000, 118000},
 			},
 			&pb.Entity{
 				Id:          "Cloud Dataflow",
-				Aggregation: []int32{1200, 1500},
+				Aggregation: []float64{1200, 1500},
 			},
 			&pb.Entity{
 				Id:          "Cloud Storage",
-				Aggregation: []int32{800, 500},
+				Aggregation: []float64{800, 500},
 			},
 		},
 	}
@@ -628,7 +628,7 @@ func MockAlerts() []*pb.Entity{
 		Type: "KubernetesMigrationAlert",
 		StartDate: time.Now().AddDate(0, 0, -30).Format(types.DEFAULT_DATE_FORMAT),
 		EndDate: time.Now().Format(types.DEFAULT_DATE_FORMAT),
-		Aggregation: []int32{60000, 120000},
+		Aggregation: []float64{60000, 120000},
 		Change: &pb.ChangeStatistic{
 			Ratio:  0,
 			Amount: 0,
@@ -636,7 +636,7 @@ func MockAlerts() []*pb.Entity{
 		Services: []*pb.Entity{
 			&pb.Entity{
 				Id:          "service-a",
-				Aggregation: []int32{20000, 10000},
+				Aggregation: []float64{20000, 10000},
 				Change: &pb.ChangeStatistic{
 					Ratio:  -0.5,
 					Amount: -10000,
@@ -645,7 +645,7 @@ func MockAlerts() []*pb.Entity{
 			},
 			&pb.Entity{
 				Id:          "service-b",
-				Aggregation: []int32{30000, 15000},
+				Aggregation: []float64{30000, 15000},
 				Change: &pb.ChangeStatistic{
 					Ratio:  -0.5,
 					Amount: -15000,
